editing my models to reflect the calls I will need to make from the API as well as creating my api call.

+++ CampMa8/Models/CampExperience.cs
@@ -11,6 +11,6 @@ namespace CampMa8.Models
         [Key]
         public int Id { get; set; }
         [Display(Name = "Camp Experience")]
-        public int Experience { get; set; }
+        public double Experience { get; set; }
     }
 }
